Fixed a bug with default selection and added support for hiding state column
NSTreePopUpButton always showed the first item even if the connected
NSTreeController didn’t have any selection. This is fixed now. If
NSTreeController doesn’t have a selection the popup button shows an
empty cell. Setting the NSMenu in Interface Builder (or by code) to
hide the state columns did not work. This is fixed now, too.